Fix some glitches in documentation (typos, co wrapper API, missing yields)
The `co` wrapper doesn't take the `onSuccess` and `onError` args, but rather
returns a Promise that can either resolve or reject. The callbacks would be
passed as args to the generator function, and that isn't the intent.

Adds `yield` to the `race` call examples that wouldn't work without it.

Fixes a few typos (my spellchecker did all the work and highlighted them)