fork control: the first arg is not iterator, but generator function: update docs and arg name
The README contains an example where the `fork` control gets an iterator (created by calling a
generator function) as the first arg. That's not what the control expects: the arg is supposed
to be a generator function that returns an iterator only after it's called.

This PR updates the README, and also renames the internal field from the misleading `iterator`
to the correct name: `generator`.